+++ archive.php
@@ -5,7 +5,7 @@
 		<div class="row">
 			<div class="<?=FULLWIDTH?>">
     			<?php the_archive_title('<h1>','</h1>'); ?>
-				<?php get_template_part('content', 'archive'); ?>
+				<?php get_template_part('template-parts/content', 'archive'); ?>
 			</div>
 		</div>
 	</div>