+++ home.php
@@ -5,7 +5,7 @@
 		<div class="row">
 			<div class="<?=FULLWIDTH?>">
     			<h1><?=__('Blog','sherpa')?></h1>
-				<?php get_template_part('content', 'archive'); ?>
+				<?php get_template_part('template-parts/content', 'archive'); ?>
 			</div>
 		</div>
 	</div>