+++ page-home.php
@@ -8,7 +8,7 @@
 	<div class="container">
 		<div class="row">
 			<div class="<?=FULLWIDTH?>">
-				<?php get_template_part('content', 'page-home'); ?>
+				<?php get_template_part('template-parts/content', 'page-home'); ?>
 			</div>
 		</div>
 	</div>