+++ page-testimonials.php
@@ -11,7 +11,7 @@
 			<div class="<?=FULLWIDTH?>">
     			<article>
         			<h1><?=__('Testimonials')?></h1>
-    				<?php get_template_part('content', 'testimonials'); ?>
+    				<?php get_template_part('template-parts/content', 'testimonials'); ?>
     			</article>
 			</div>
 		</div>