+++ template-parts/content-page-home.php
@@ -1,0 +1,12 @@
+<?php
+
+	if(have_posts()) {
+		while(have_posts()) {
+			the_post();
+			
+			echo "<article>";
+			the_content();
+			echo "</article>";
+		}
+		
+	}